We now uses the api gateway to communicate between auth and user service. Other services is in wip

+++ TwitterMock/UserService/Controllers/UserController.cs
@@ -1,3 +1,5 @@
+using System.Text;
+using System.Text.Json;
 using Microsoft.AspNetCore.Mvc;
 using UserService.Core.Services.Dtos;
 using UserService.Core.Services.Interfaces;
@@ -8,14 +10,14 @@ namespace UserService.Controllers;
 [Route("api/[controller]")]
 public class UserController : ControllerBase
 {
-    
     private readonly IUserService _userService;
+    private readonly HttpClient _client = new HttpClient();
 
     public UserController(IUserService userService)
     {
         _userService = userService;
     }
-    
+
     [HttpPost]
     [Route("AddUser")]
     public async Task<IActionResult> AddUser([FromBody] CreateUserDTO dto)
@@ -30,13 +32,18 @@ public class UserController : ControllerBase
             return BadRequest(e.Message);
         }
     }
-    
+
     [HttpGet]
     [Route("GetUsers")]
-    public async Task<IActionResult> GetUsers([FromQuery] PaginatedDTO paginatedDto)
+    public async Task<IActionResult> GetUsers([FromQuery] PaginatedDTO paginatedDto, [FromHeader] string token)
     {
         try
         {
+            var url = "http://localhost:5206/api/Auth/ValidateToken";
+            _client.DefaultRequestHeaders.Add("token", token);
+            var result = await _client.GetAsync(url);
+
+            if (!result.IsSuccessStatusCode) return Unauthorized(result.RequestMessage);
             return Ok(await _userService.GetUsers(paginatedDto));
         }
         catch (Exception e)
@@ -44,27 +51,42 @@ public class UserController : ControllerBase
             return BadRequest(e.Message);
         }
     }
-    
+
     [HttpGet]
     [Route("{userId}")]
-    public async Task<IActionResult> GetUserById([FromRoute] int userId)
+    public async Task<IActionResult> GetUserById([FromRoute] int userId, [FromHeader] string token)
     {
         try
         {
-            return Ok(await _userService.GetUserById(userId));
+            var url = "http://localhost:5206/api/Auth/ValidateToken";
+            _client.DefaultRequestHeaders.Add("token", token);
+            var result = await _client.GetAsync(url);
+
+            if (result.IsSuccessStatusCode)
+            {
+                return Ok(await _userService.GetUserById(userId));
+            }
+
+            return Unauthorized();
         }
         catch (Exception e)
         {
             return BadRequest(e.Message);
         }
     }
-    
+
     [HttpPut]
     [Route("{userId}")]
-    public async Task<IActionResult> UpdateUser([FromRoute] int userId, [FromBody] UpdateUserDTO dto)
+    public async Task<IActionResult> UpdateUser([FromRoute] int userId, [FromBody] UpdateUserDTO dto,
+        [FromHeader] string token)
     {
         try
         {
+            var url = "http://localhost:5206/api/Auth/ValidateToken";
+            _client.DefaultRequestHeaders.Add("token", token);
+            var result = await _client.GetAsync(url);
+
+            if (!result.IsSuccessStatusCode) return Unauthorized();
             await _userService.UpdateUser(userId, dto);
             return Ok();
         }
@@ -73,14 +95,19 @@ public class UserController : ControllerBase
             return BadRequest(e.Message);
         }
     }
-    
-    
+
+
     [HttpDelete]
     [Route("{userId}")]
-    public async Task<IActionResult> DeleteUser([FromRoute] int userId)
+    public async Task<IActionResult> DeleteUser([FromRoute] int userId, [FromHeader] string token)
     {
         try
         {
+            var url = "http://localhost:5206/api/Auth/ValidateToken";
+            _client.DefaultRequestHeaders.Add("token", token);
+            var result = await _client.GetAsync(url);
+
+            if (!result.IsSuccessStatusCode) return Unauthorized();
             await _userService.DeleteUser(userId);
             return Ok();
         }
@@ -89,5 +116,4 @@ public class UserController : ControllerBase
             return BadRequest(e.Message);
         }
     }
-
 }
